:sparkles: Filter to avoid send empty sheets to vendor added
In order to send just relevant data to vendors, a new filter was applied to explicity delete a vendor from list of the ones to send emails when all his purchase orders are valid based on Firebase data. This has no relationship whit FUP sheet data, that means if for a vendor there is data in FUP sheet, but in Firebase that data is fine, any purchase order data request will be send to it. Also structure to deployed files was modified in order to improve readability